Se pudo conectar la consola del patron Redux
