+++ package.json
@@ -6,8 +6,10 @@
     "bootstrap": "^4.3.1",
     "react": "16.8.6",
     "react-dom": "16.8.6",
+    "react-redux": "^7.2.1",
     "react-router-dom": "5.0.0",
-    "react-scripts": "3.0.0"
+    "react-scripts": "3.0.0",
+    "redux": "^4.0.5"
   },
   "scripts": {
     "start": "react-scripts start",
