+++ src/reducers/index.js
@@ -1,4 +1,4 @@
-import { combineReducers } from 'react-redux'
+import { combineReducers } from 'redux'
 import buscadorImagenReducer from '../reducers/buscadorImagenReducer'
 
 export default combineReducers({
